+++ backend/src/main/java/com/trick/backend/model/dto/RechargeDTO.java
@@ -4,9 +4,11 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
+
 @Data
 @AllArgsConstructor
 @NoArgsConstructor
 public class RechargeDTO {
-    Double amount;
+    BigDecimal amount;
 }
