+++ backend/src/main/java/com/trick/backend/model/vo/ChargingPileVO.java
@@ -4,6 +4,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -15,8 +16,8 @@ public class ChargingPileVO {
     private String pileNo;
     private Integer type;
     private Integer status;
-    private Double powerRate;
-    private Double pricePerKwh;
+    private BigDecimal powerRate;
+    private BigDecimal pricePerKwh;
     private String locationDesc;
     private Double latitude;
     private Double longitude;
