+++ backend/src/main/java/com/trick/backend/service/TransactionLogService.java
@@ -3,10 +3,12 @@ package com.trick.backend.service;
 import com.trick.backend.common.result.PageResult;
 import com.trick.backend.model.pojo.TransactionLog;
 
+import java.math.BigDecimal;
+
 public interface TransactionLogService {
 
     PageResult<TransactionLog> getPagedTransactions(Integer id, Integer pageNum, Integer pageSize);
 
-    void recharge(Integer id, Double amount);
+    void recharge(Integer id, BigDecimal amount);
 
 }
